Recursively iterate over a specified directory and return file paths.
Right now this will only work when a folder called 'root' is found in
the save directory. We will recursively iterate over all folders and
their subfolders and return the full path of each file (empty folders
are ignored) using recursivelyGetDirectoryItems(...). This function
returns a sequence containing the paths.

We then use this sequence to create a tree. Each folder is a node in the
tree and will have links to his children (= subfolders) and the files it
contains.