Close #4 - Use patterns in the config file to ignore certain files.
LoFiVi now creates a config file in the save directory. In this file the
user can use patterns or simple strings to ignore certain files and
folders when a graph is created.

The default setting ignores files and folders that start with a fullstop
(e.g.: .git, .hg, .DS_Store, etc.).