Use pre-calculated layout for files around folder nodes.
Files aren't affected by physical forces anymore. Instead, we calculate
a uniform layout based on the amount of files around the node. Files
are placed on "onion layers" around a folder node. A new layer is
created once the current layer becomes too crowded (we check the
distance between each file on the layer and see if it is lower than a
certain threshold).

One of the benefits is increased performance of course, since we no
longer need to calculate physical forces between the files.

It also reduces the initial "chaos" when the graph is created and
looking for an equilibrium.

Last but not least it improves the layout of the files, increases the
readability of the graph and makes it adjustable.