Fix #6 - Correctly log files without extension.
After splitting the path, the function looked for a fullstop to identify
a file. This of course only worked when the file had an extension. Files
without an extension didn't have a fullstop in their name and therefore
were logged as a folder.

We now use a second loop to determine when we are looking at the last
part of the the split path (which has to be a file), instead of looking
for a fullstop.