Postpone graph creation
When LoFiVi starts it will no longer create a graph for the files inside
of the root folder. Instead, it will display a info panel telling the
user to drop a folder on the application.

Once that folder is dropped onto LoFiVi it will be used to create a
graph.